Add option to add extra ignored prefixes (#26)
We add an option to extend `IGNORE_PREFIXES` with extra prefixes we may need
to ignore.

Some web archives may load additional resources (e.g. web stats or a
toolbar) in the playback page and wombat tries to rewrite these URLs as
well. We need a way to skip them.

Co-authored-by: Vangelis Banos <vangelis@archive.org>