WIP: make tests work on macOS
This replaces `just-launch-chrome` with `chrome-launcher` and tweaks some of the internals. With this change, all of the tests can find, launch, and then close the browser, and _some_ of them pass. However, the majority of tests still fail on my mac. Here's a copy of the logs: https://gist.github.com/nfriedly/2917ff09e7560cb0ea54694ca525de49

I initially tried to stick with something closer to the just-launch-chrome API

```js
   });
+  const fetch = require('node-fetch');
+  const details = await fetch(`http://localhost:${chrome.port}/json`).then(res => res.json());
+  const browserWSEndpoint = details[0].webSocketDebuggerUrl;
   const browser = await Browser.connect(browserWSEndpoint, {
```

But, it didn't seem to be able to connect to the browser. The `details[0].webSocketDebuggerUrl` was definitely a websocket URL, but perhaps it wasn't the right one.

When complete, this should fix #44